revert: back to int8 baseline + keep buffer pool (round-1 known-good)
Three int16 iterations (rounds 2/2.1/2.2) all OOM-killed under load,
producing scores 805-1180 with 281-413 HTTP errors. Even after VP-tree
compaction (16→13 bytes/node), the int16 vectors at 84 MB plus V8
baseline overhead pushed RSS past 165 MB cap in CI.

Decision: revert quantization/heap/tree code to round-1 (commit de168f9)
int8 baseline. Keep two genuine improvements from later rounds:
  - buffer pool in api.ts (zero per-request Buffer.concat)
  - LB CPU 0.20, API 0.40 (the proven CPU split — 0.10/0.45 starves LB)

Reverts:
  src/quantize.ts:  back to int8 (scale 127, sentinel -128)
  src/vptree.ts:    back to 16-byte node format (Float32 + 3×Int32)
  src/heap.ts:      back to Int32Array dists
  src/build-index.ts: emits int8 vectors
  src/api.ts:       Int16Array views → Int8Array views
  src/test/*.ts:    Int16Array → Int8Array

Dockerfile.api: NODE_OPTIONS heap 16 → 64 (no need to be tight on int8;
keep room for V8 GC under load).

docker-compose.yml: pinned image :int8-stable, mem 165→160 per API,
LB 20→30 MB.

Local verify-detection matches round-1 result exactly:
  TP=23998 TN=29991 FP=51 FN=60 = 0.205%

+++ src/test/smoke.ts
@@ -44,7 +44,7 @@ async function main(): Promise<void> {
   console.log(`[smoke] refs=${refs.length} payloads=${payloads.length}`);
 
   const N = refs.length;
-  const vectors = new Int16Array(N * DIMS);
+  const vectors = new Int8Array(N * DIMS);
   const labels = new Uint8Array(Math.ceil(N / 8));
 
   for (let i = 0; i < N; i++) {
@@ -66,7 +66,7 @@ async function main(): Promise<void> {
   assertEq(tree2.size, tree.size, 'serialize round-trip size');
 
   const queryFloat = new Float32Array(DIMS);
-  const queryInt8 = new Int16Array(DIMS);
+  const queryInt8 = new Int8Array(DIMS);
   const heap = new TopKHeap(5);
 
   let queriesRun = 0;
@@ -96,8 +96,8 @@ async function main(): Promise<void> {
 }
 
 function bruteForceTop5(
-  vectors: Int16Array,
-  query: Int16Array,
+  vectors: Int8Array,
+  query: Int8Array,
   N: number,
 ): Array<{ d: number; i: number }> {
   const all: Array<{ d: number; i: number }> = new Array(N);
